+++ custom_components/mydolphin_plus/managers/config_manager.py
@@ -7,6 +7,8 @@ from cryptography.fernet import Fernet, InvalidToken
 from homeassistant.config_entries import STORAGE_VERSION, ConfigEntry
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers import translation
+from homeassistant.helpers.entity import DeviceInfo, EntityDescription
 from homeassistant.helpers.json import JSONEncoder
 from homeassistant.helpers.storage import Store
 
@@ -32,6 +34,7 @@ class ConfigManager:
     _store: Store | None
     _store_data: dict | None
     _entry_data: dict | None
+    _translations: dict | None
     _password: str | None
     _entry_title: str
     _entry_id: str
@@ -53,6 +56,7 @@ class ConfigManager:
         self._store = None
         self._entry_data = None
         self._store_data = None
+        self._translations = None
 
         self._is_set_up_mode = entry is None
         self._is_initialized = False
@@ -142,6 +146,10 @@ class ConfigManager:
             self._data[CONF_USERNAME] = self._entry_data.get(CONF_USERNAME)
             self._data[CONF_PASSWORD] = password
 
+            self._translations = await translation.async_get_translations(
+                self._hass, self._hass.config.language, "entity", {DOMAIN}
+            )
+
             self._is_initialized = True
 
         except InvalidToken:
@@ -160,6 +168,21 @@ class ConfigManager:
             _LOGGER.error(
                 f"Failed to initialize configuration manager, Error: {ex}, Line: {line_number}"
             )
+
+    def get_entity_name(
+        self, platform, entity_description: EntityDescription, device_info: DeviceInfo
+    ) -> str:
+        entity_key = entity_description.key
+
+        device_name = device_info.get("name")
+
+        translation_key = f"component.{DOMAIN}.entity.{platform}.{entity_key}.name"
+
+        translated_name = self._translations.get(translation_key, device_name)
+
+        entity_name = f"{device_name} {translated_name}"
+
+        return entity_name
 
     def update_credentials(self, data: dict):
         self._entry_data = data